// Front-end/utils/type.ts
// export an enum called MessageType with the following values
export enum MessageType {
  // info message type
  info = "info",
  // success message type
  success = "success",
  // error message type
  error = "error",
  // warning message type
  warning = "warning",
  // loading message type
  loading = "loading",
  // warn message type
  // warn = "warn",
}